# .well-known/openapi.yaml
openapi: 3.1.0
info:
  title: Mario Personal MCP API
  version: 1.0.0
  description: Public API exposing Mario Mayerle's LinkedIn profile, GitHub repositories, blog posts, and indexed books.
servers:
  - url: https://mcp.mariomayerle.com
paths:
  /resources/livros:
    post:
      summary: Query insights from Mario's books
      operationId: queryBooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                  description: User's question about the book content
      responses:
        '200':
          description: Response retrieved from Mario's book collection
...
  /tools/get_project_details:
    get:
      summary: Retrieve details of a specific GitHub repository
      operationId: getProjectDetails
      parameters:
        - in: query
          name: repo_name
          required: true
          schema:
            type: string
          description: Exact name of the repository to query
      responses:
        '200':
          description: Repository details retrieved successfully
        '404':
          description: Repository not found